[UPDATE]
Syscalls completadas, inicio de documentación.

+++ Pruebas/pruebas3p/taro_get.c
@@ -8,7 +8,7 @@
 
 struct memory_limitation {
     pid_t pid;
-    long limit;
+    size_t limit;
 };
 
 int main() {
